+++ packages/annotation-converter/src/converter.ts
@@ -1355,6 +1355,16 @@ function convertNavigationProperty(
         if (refConstraints) {
             convertedNavigationProperty.referentialConstraint =
                 (refConstraints.collection?.map((record) => {
+                    const referencedProperty = (record as AnnotationRecord).propertyValues.find((prop) => {
+                        return prop.name === 'ReferencedProperty';
+                    })?.value as { PropertyPath?: string };
+                    let targetedProperty = '';
+                    if (referencedProperty.PropertyPath) {
+                        const targetedPropertySplit = referencedProperty.PropertyPath.split('/');
+                        targetedPropertySplit.shift();
+                        targetedProperty = targetedPropertySplit.join('/');
+                    }
+
                     return {
                         sourceProperty: (
                             (record as AnnotationRecord).propertyValues.find((prop) => {
@@ -1362,11 +1372,7 @@ function convertNavigationProperty(
                             })?.value as { PropertyPath?: string }
                         ).PropertyPath,
                         targetTypeName: convertedNavigationProperty.targetTypeName,
-                        targetProperty: (
-                            (record as AnnotationRecord).propertyValues.find((prop) => {
-                                return prop.name === 'ReferencedProperty';
-                            })?.value as { PropertyPath?: string }
-                        ).PropertyPath
+                        targetProperty: targetedProperty
                     };
                 }) as any) ?? [];
         }
@@ -1439,6 +1445,16 @@ function convertAction(converter: Converter, rawAction: RawAction): Action {
 
     if (convertedAction.returnType) {
         lazy(convertedAction, 'returnEntityType', resolveEntityType(converter, rawAction.returnType));
+        lazy(convertedAction, 'returnTypeReference', () => {
+            const typeName = convertedAction.returnType.startsWith('Collection')
+                ? convertedAction.returnType.substring(11, convertedAction.returnType.length - 1)
+                : convertedAction.returnType;
+            return (
+                converter.getConvertedEntityType(typeName) ??
+                converter.getConvertedComplexType(typeName) ??
+                converter.getConvertedTypeDefinition(typeName)
+            );
+        });
     }
 
     lazy(convertedAction, 'parameters', converter.convert(rawAction.parameters, convertActionParameter));
@@ -1495,14 +1511,17 @@ function convertActionParameter(
 ): ActionParameter {
     const convertedActionParameter = rawActionParameter as ActionParameter;
 
-    lazy(
-        convertedActionParameter,
-        'typeReference',
-        () =>
-            converter.getConvertedEntityType(rawActionParameter.type) ??
-            converter.getConvertedComplexType(rawActionParameter.type) ??
-            converter.getConvertedTypeDefinition(rawActionParameter.type)
-    );
+    lazy(convertedActionParameter, 'typeReference', () => {
+        let targetType = rawActionParameter.type;
+        if (targetType.startsWith('Collection(')) {
+            targetType = targetType.substring(11, targetType.length - 1);
+        }
+        return (
+            converter.getConvertedEntityType(targetType) ??
+            converter.getConvertedComplexType(targetType) ??
+            converter.getConvertedTypeDefinition(targetType)
+        );
+    });
 
     lazy(convertedActionParameter, 'annotations', () => {
         // annotations on action parameters are resolved following the rules for actions