+++ packages/fe-mockserver-admin/src/mockdata/EntityContainer.ts
@@ -1,10 +1,14 @@
-import type { MockEntityContainerContributor, ODataRequest } from '@sap-ux/fe-mockserver-core';
+import type { ODataRequest } from '@sap-ux/fe-mockserver-core';
+import { MockEntityContainerContributorClass } from '@sap-ux/fe-mockserver-core';
 
-const MockEntityContainer: MockEntityContainerContributor = {
+export default class MockEntityContainer extends MockEntityContainerContributorClass {
     async handleRequest(odataRequest: ODataRequest): Promise<unknown> {
+        this.myMethod();
+        // Example of handling a request and returning a custom response
         odataRequest.addResponseHeader('content-type', 'text/html ;charset=utf-8');
         return `<html><body><h1>Hello</h1></body></html>`;
     }
-};
-
-export default MockEntityContainer;
+    myMethod(): void {
+        this.base.getEntityInterface('eoeoe');
+    }
+}