+++ packages/fe-mockserver-admin/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @sap-ux/fe-mockserver-admin
 
+## 0.0.14
+
+### Patch Changes
+
+-   Updated dependencies [bf2320d]
+    -   @sap-ux/fe-mockserver-core@1.6.18
+
+## 0.0.13
+
+### Patch Changes
+
+-   Updated dependencies [e3538f5]
+    -   @sap-ux/fe-mockserver-core@1.6.17
+
+## 0.0.12
+
+### Patch Changes
+
+-   Updated dependencies [2ede21e]
+    -   @sap-ux/fe-mockserver-core@1.6.16
+
+## 0.0.11
+
+### Patch Changes
+
+-   Updated dependencies [85ce46a]
+    -   @sap-ux/fe-mockserver-core@1.6.15
+
+## 0.0.10
+
+### Patch Changes
+
+-   Updated dependencies [3636693]
+    -   @sap-ux/fe-mockserver-core@1.6.14
+
+## 0.0.9
+
+### Patch Changes
+
+-   Updated dependencies [a953dd9]
+    -   @sap-ux/fe-mockserver-core@1.6.13
+
+## 0.0.8
+
+### Patch Changes
+
+-   Updated dependencies [eb7176f]
+    -   @sap-ux/annotation-converter@0.10.7
+    -   @sap-ux/fe-mockserver-core@1.6.12
+
 ## 0.0.7
 
 ### Patch Changes